chore: update typescript and refactor solution
