+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "imgstry",
   "version": "1.0.0",
-  "description": "An image editing module written in JavaScript.",
+  "description": "A JavaScript image editing module.",
   "main": "index.js",
   "directories": {
     "example": "example"
@@ -29,27 +29,32 @@
   },
   "homepage": "https://github.com/vladjerca/imgstry#readme",
   "devDependencies": {
-    "browser-sync": "^2.12.8",
-    "browserify": "^13.0.1",
-    "del": "^2.2.0",
-    "chai": "^3.5.0",
-    "gulp": "^3.9.1",
-    "gulp-foreach": "^0.1.0",
-    "gulp-istanbul": "^0.10.4",
-    "mochawesome": "^1.3.5",
-    "gulp-mocha": "^2.2.0",
-    "gulp-mocha-phantomjs": "^0.11.0",
-    "should": "^9.0.2",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-tslint": "^5.0.0",
-    "gulp-typescript": "^2.13.4",
-    "gulp-uglify": "^1.5.3",
-    "mocha": "^2.5.3",
-    "merge2": "^1.0.2" ,
-    "run-sequence": "^1.2.1",
-    "tslint": "^3.10.2",
-    "typescript": "^1.8.10",
-    "vinyl-buffer": "^1.0.0",
-    "vinyl-source-stream": "^1.1.0"
+    "@types/assertion-error": "1.0.30",
+    "@types/chai": "3.4.34",
+    "@types/mocha": "2.2.32",
+    "@types/should": "8.1.30",
+    "@types/sinon": "1.16.31",
+    "browser-sync": "2.17.1",
+    "browserify": "13.1.0",
+    "chai": "3.5.0",
+    "del": "2.2.2",
+    "gulp": "3.9.1",
+    "gulp-foreach": "0.1.0",
+    "gulp-istanbul": "1.1.1",
+    "gulp-mocha": "3.0.1",
+    "gulp-mocha-phantomjs": "0.12.0",
+    "gulp-sourcemaps": "1.6.0",
+    "gulp-tslint": "6.1.2",
+    "gulp-typescript": "3.0.2",
+    "gulp-uglify": "2.0.0",
+    "merge2": "1.0.2",
+    "mocha": "3.1.0",
+    "mochawesome": "1.5.3",
+    "run-sequence": "1.2.2",
+    "should": "11.1.0",
+    "tslint": "3.15.1",
+    "typescript": "2.0.3",
+    "vinyl-buffer": "1.0.0",
+    "vinyl-source-stream": "1.1.0"
   }
 }
