+++ source/pixel/color/spaces/rgb.ts
@@ -1,7 +1,9 @@
-/// <reference path="./interfaces/color.d.ts" />
-import { Hsv } from './hsv';
-import { Cmyk } from './cmyk';
-import { Hex } from './hex';
+import {
+  Color,
+  Hsv,
+  Cmyk,
+  Hex,
+ } from '../../color';
 
 export class Rgb implements Color {
   public r: number;
