+++ source/pixel/index.ts
@@ -1,0 +1,2 @@
+export * from './color';
+export * from './pixel';
