+++ test/integration/cmyk.test.ts
@@ -1,8 +1,6 @@
-/// <reference path="../../typings/chai/chai.d.ts" />
-/// <reference path="../../typings/mocha/mocha.d.ts" />
 import { expect } from 'chai';
-import { colorMap } from './colors';
-import { Cmyk } from '../../source/pixel/color/cmyk';
+import { colorMap } from './constants/colors';
+import { Cmyk } from '../../source/pixel';
 
 describe('CMYK color', () => {
   it('Should have all channels 0 initially', () => {
