+++ test/integration/hsv.test.ts
@@ -1,9 +1,7 @@
-/// <reference path="../../typings/chai/chai.d.ts" />
-/// <reference path="../../typings/mocha/mocha.d.ts" />
 import { expect } from 'chai';
-import { colorMap } from './colors';
-import { Hsv } from '../../source/pixel/color/hsv';
-import { Hex }  from '../../source/pixel/color/hex';
+import { colorMap } from './constants/colors';
+import { Hsv } from '../../source/pixel';
+import { Hex }  from '../../source/pixel';
 
 describe('HSV color', () => {
   it('Should have all channels 0 initially', () => {
