+++ test/integration/rgb.test.ts
@@ -1,9 +1,7 @@
-/// <reference path="../../typings/chai/chai.d.ts" />
-/// <reference path="../../typings/mocha/mocha.d.ts" />
 import { expect } from 'chai';
-import { colorMap } from './colors';
-import { Rgb } from '../../source/pixel/color/rgb';
-import { Hex } from '../../source/pixel/color/hex';
+import { colorMap } from './constants/colors';
+import { Rgb } from '../../source/pixel';
+import { Hex } from '../../source/pixel';
 
 describe('RGB color', () => {
   it('Should have all channels 0 initially', () => {
