--- tsd.json
@@ -1,18 +1,0 @@
-{
-  "version": "v4",
-  "repo": "borisyankov/DefinitelyTyped",
-  "ref": "master",
-  "path": "typings",
-  "bundle": "typings/tsd.d.ts",
-  "installed": {
-    "mocha/mocha.d.ts": {
-      "commit": "b9989e602b3225d48aa4a9dfa0631ae41fea100d"
-    },
-    "assertion-error/assertion-error.d.ts": {
-      "commit": "7ad4e09016898d9c4837e250121fef2aedbecaaa"
-    },
-    "chai/chai.d.ts": {
-      "commit": "7ad4e09016898d9c4837e250121fef2aedbecaaa"
-    }
-  }
-}
