--- typings/assertion-error/assertion-error.d.ts
@@ -1,15 +1,0 @@
-// Type definitions for assertion-error 1.0.0
-// Project: https://github.com/chaijs/assertion-error
-// Definitions by: Bart van der Schoor <https://github.com/Bartvds>
-// Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
-
-declare module 'assertion-error' {
-	class AssertionError implements Error {
-		constructor(message: string, props?: any, ssf?: Function);
-		name: string;
-		message: string;
-		showDiff: boolean;
-		stack: string;
-	}
-	export = AssertionError;
-}
