--- typings/index.d.ts
@@ -1,4 +1,0 @@
-/// <reference path="globals/parallel/index.d.ts" />
-/// <reference path="globals/should/index.d.ts" />
-/// <reference path="modules/chai/index.d.ts" />
-/// <reference path="modules/sinon/index.d.ts" />
