--- typings/modules/chai/typings.json
@@ -1,21 +1,0 @@
-{
-  "resolution": "main",
-  "tree": {
-    "src": "https://raw.githubusercontent.com/typed-typings/npm-chai/793bee097a6a644e078a033603d88ac89eb7b560/typings.json",
-    "raw": "registry:npm/chai#3.5.0+20160415060238",
-    "main": "index.d.ts",
-    "global": false,
-    "dependencies": {
-      "assertion-error": {
-        "src": "https://raw.githubusercontent.com/typed-typings/npm-assertion-error/105841317bd2bdd5d110bfb763e49e482a77230d/typings.json",
-        "raw": "github:typed-typings/npm-assertion-error#105841317bd2bdd5d110bfb763e49e482a77230d",
-        "main": "main.d.ts",
-        "global": false,
-        "name": "assertion-error",
-        "type": "typings"
-      }
-    },
-    "name": "chai",
-    "type": "typings"
-  }
-}
