--- typings/modules/sinon/typings.json
@@ -1,11 +1,0 @@
-{
-  "resolution": "main",
-  "tree": {
-    "src": "https://raw.githubusercontent.com/typed-typings/npm-sinon/e18b1896080e7ae16c16a2725ed5e43d2f2fea65/typings.json",
-    "raw": "registry:npm/sinon#1.16.0+20160427193336",
-    "main": "lib/sinon.d.ts",
-    "global": false,
-    "name": "sinon",
-    "type": "typings"
-  }
-}
