--- typings/tsd.d.ts
@@ -1,3 +1,0 @@
-/// <reference path="mocha/mocha.d.ts" />
-/// <reference path="assertion-error/assertion-error.d.ts" />
-/// <reference path="chai/chai.d.ts" />
